math/nbre: add methods letting user know if a number is in/de-creasing or not.
Adding a method letting user of a number is a increasing one (its digits
are increasing).
Adding an another method letting user of a number is a decreasing one
(its digits are decreasing).
Add a third one letting user know if a number is a bouncy one (neither
an increasing one, neither a decreasing one).