math/nbre: fix a crash.
The method 'build_number_from_digits_rev' lets user build a number from an
array of digits. The suffix 'rev' indicates that the number will be
built from the last digit of the array to the first.
For this purpose, the index starts from last_idx and stops when become
below than first_idx. But, as the index variable is unsigned, if first_idx is
equal to 0, it will crash.
Fix that.