math/nbre: defines the number as long instead of int.
Several methods in these files still take number from whom the work is
done as simple int. We should take long to avoid issues.
Project Euler will often ask to use big variables and it will not take too much
time to use a long instead of a int when just an int is necessary is the
calling method.