macros: add 'RETHROW' macros.
The MACRO RETHROW exit the current method if the arg is < 0 and return
arg.
The MACRO RETHROW_P exit the current method if the arg is NULL and return
arg.
The MACRO RETHROW_PN exit the current method if the arg is NULL and
return -1.
The MACRO RETHROW_NP exit the current method if the arg is < 0 and
return NULL.