src/liste: unify method 'gl_add_elem_trie_*'.
The method 'gl_add_elem_trie' took an argument 'sens_croissant' that
indicated in which order the list should be sorted.
And then, it called a method 'gl_add_elem_trie_c" for sorting te list in
the increasing order and 'gl_add_elem_trie_d' for the decreasing order.

But, it is useless to do that. The order will be decided by the 'cmp_elem'
method that sorts the elements.