src/liste: rename 'gl_remove_*' to 'gl_delete_*'.
I am implementing new methods letting take an element from a list
without deleting it. But, it could be confusing with the methods
'gl_remove_*'. We could think these methods take an element from the
list without deleting it too.
Rename these methods as 'gl_delete_*' in order to have a clearer code.