src/liste: change nbre_elem list signed.
list.nbre_elem was unsigned int. It can appear logical as an list can
not have negatives elements.
But, it is very annoying as if an user wants to get the element N, it
has to give this N to the method but we need to check the value. As
comparison between signed and unsigned are not possible, check that the
value given is before 0 is not possible...