src/liste: update 'gl_contains_data' signature.
The method 'gl_contains_data' takes a no const liste because it uses the
method 'gl_get_elem_data' that returns a no-const gl_elem_t *.
But, it is just better, according to me, to give a const list * in the
method 'gl_contains_data'.
Update its signature and redo the very small code done in
'gl_get_elem_data' in order to know if the element is present or not.