tests: use the new methods defined in 'test_helper'.
Tests "framework" has been completely redone.
Now, a list with all tests methods is built and the,n these tests are
run. The point of building a list is filter the module or the test to
run. This is done like this:
'./test [-m MODULE |-t TEST]]'.

The argument 'm' and 't' are optional and 'm' option is necessary if the
user wants to indicate 't' option.

The option 'm' lets run all the tests of a module. The option 't' lets
run only one test in this module.