src/vector: add the macros 'gv_wipe' and 'gv_clear'.
'gv_wipe' calls the method '_gv_wipe' (giving it the size of the
element). It frees all the elements of the tab if 'free_data_cb' has
been provided by the user and then free the tab.
It should be used for a vector no dynamically allocated when the user
does not need this vector anymore.

The macros 'gv_clear' calls the method '_gv_clear' (giving it the size
of the element) and clear the tab but does not deallocate it. So, the
vector can still be used and the fact not deallocated the tab lets user
not allocate new tab when adding new elements.