src/vector: add the 'size_elem' field and set it.
On vector, we regularly work with the "void vector": every time, a
"true" method is called with a vector as argument.
In these methods, we need to know the size of the element.
This one can be indicated from the macro with '__gv_size'. That works
but I do not like the fact to be forced to always call this macro
whereas the size can be known.
Indeed, when the vector is initialized (via the macro 'gv_init'), we
could set a field 'size_t' containing the size of the element.
As this size will never change for the current vector, the size will be
then available for the rest of the life of the vector.

However, we need to guarantee that this size will never be modified.
So, the field is defined as constant. But, on initialization, we need to
modify this field. It's the only time we need to modify it.
So, the macro 'unconst_cast' is used.