src/vector: update the macro 'gv_init'.
Return the pointer of the vector initialized in thre macro 'gv_init'.
As, in the macro 'gv_new', after having allocated the vector via the
'p_malloc', we need to initialize this vector ('p_calloc' does not
complete the job as we need to set the field '__size_elem', return the
pointer of the vector in the macro 'gv_init' in order to not stock it in
'gv_new', calls 'gv_init' and returns it.