src/vector: add a way to remove elements.
Add the macro 'gv_remove' and the associated method '__gv_remove'.
This method does not really remove the element. It decreases the length
of the vector and if this element in in the vector, use 'memmove' in
order to move the data after the element to remove from one block to the
beginning.
So, if the vector has this:
'1, 2, 3, 4' with len and size equal to 4 and he removes the user
removes the element 3, the vector will then has this:
'1, 2, 4' but the memory space on vec->tab[3] will be still available
and still contain '4'.