src/vector: fix issue in 'gv_init_size'.
The macro 'gv_init_size' lets user initialize the vector with a given
size to the vector. So, the array will be allocated with the given size.
That can let user win some time after as the array will not be
reallocated each time a new element is added (until thr array is
filled).
But, the macro called '__gv_grow' that increases the length of the
vector too. Fix that issue calling rather '__gv_extend'.

Add a test about this issue too.