src/vector: improve the macro 'gv_insert_elem_sorted'.
This macro lets user insert a new element in the vector, after having
searched the spot in order to keep the vector sorted.
In order to search this spot in a method, we need to use the
'__vector_void_t' and so a void *data.
But, if the user wants to add a simple integer in the vector, it has to
set it in a variable and then pass the pointer to this variable, with
the actual implementation of this macro.

Improve it declaring this variable inside the macro. That lets user just
indicate the new element directly.