src/big_numbers: add the field 'limit'.
In the structure 'big_number_t' representing a big_numbers, add the
field 'limit'.
It will correspond to the maximal value a part of a big number will able
to stock. If it is overtaken, a new member in the vector (inside the
big_number) will be created and the carry set in it.