src/big_numbers: define 'limit' as a const.
The filed 'limit' will be almost never modified. This case will be very
rare. So, it should be constant.

Define it as a constant variable in the structure 'big_number_t' and add a
method in order to still have a way to update its value to change its value.